fix(mcp,core): trim MCP recall hits, render brief and lint as text, no migration backups on a fresh store
Soak gaps 6 and 7 from the 2026-08-25 usefulness pass.

- MCP `recall` (and the Brief's related context) returns the fields a skill
  reads: source_id, title, content, score, last_modified, memory_type,
  space, source_agent, confidence, confirmed, entity_name, supersedes and
  the revision fields. Chunk index, raw score, source text, content hash
  and the other ranking and ingest internals leave the agent's view.
- MCP `brief` renders the Brief as text (space, version, last session,
  active and backlog items with their ids, versions and gates) instead of
  the raw JSON.
- MCP `lint` returns the text `wenlan lint` prints as its content and keeps
  the typed report in `structured_content` for the skill's digests and
  completeness flags. The renderer moves from the CLI to
  `LintReport::render_text` in wenlan-types so both share it.
- A database created in this boot (its migration chain starts at
  `user_version` 0) takes no pre-migration backups: a fresh install replays
  every migration onto an empty store and left 29
  `pre_migration_<n>_backup.db` files (37 MB) beside the live db. A
  database that already existed still gets its restore points; the two
  restore-point drills now start from a store at 96 and at 122.
