+++ crates/wenlan-types/src/lint_report.rs
@@ -316,3 +316,244 @@ impl LintErrorResponse {
         &self.error
     }
 }
+
+impl LintReport {
+    /// The report as the text `wenlan lint` prints and the MCP `lint` tool
+    /// returns: totals, per-group counts, then every finding, advisory and
+    /// incomplete check with its recommendation and coverage. The typed
+    /// report stays the contract; this is what a person or an agent reads.
+    pub fn render_text(&self) -> String {
+        let totals = self.totals();
+        let mut groups = [(0_u32, 0_u32, 0_u32); 7];
+        for check in self.checks() {
+            let Some(group) = LintCheckGroup::for_check_id(check.check_id()) else {
+                continue;
+            };
+            let counts = &mut groups[group_index(group)];
+            counts.0 += 1;
+            match check.outcome() {
+                LintOutcome::Pass => {}
+                LintOutcome::Finding => counts.1 += 1,
+                LintOutcome::NotRunPrerequisite
+                | LintOutcome::InconsistentSnapshot
+                | LintOutcome::FailedToRun => counts.2 += 1,
+            }
+        }
+        let mut output = format!(
+            "Lint: {} checks, {} passed, {} actionable findings, {} advisor{}, {} incomplete\nGroups:\n",
+            totals.checks(),
+            totals.passed(),
+            totals.actionable_findings(),
+            totals.advisory_findings(),
+            if totals.advisory_findings() == 1 { "y" } else { "ies" },
+            totals.incomplete()
+        );
+        for group in LintCheckGroup::ALL {
+            let (checks, findings, incomplete) = groups[group_index(group)];
+            if checks == 0 {
+                continue;
+            }
+            output.push_str(&format!(
+                "  {}: {checks} check{}, {findings} findings, {incomplete} incomplete\n",
+                group.as_str(),
+                if checks == 1 { "" } else { "s" }
+            ));
+        }
+        if let Some(work) = self.agent_work() {
+            output.push_str(&format!(
+                "Agent work: {} bounded records; digest={}\n",
+                work.records().len(),
+                work.work_digest().as_str()
+            ));
+        }
+        append_findings(&mut output, "Findings", self, LintGateEffect::Actionable);
+        append_findings(&mut output, "Advisories", self, LintGateEffect::Advisory);
+        // A check that passed but still carries advice is waiting on the owner, not
+        // broken — e.g. a graph channel with no model source chosen. It is not a
+        // finding and does not change the exit code, but the reader still has to see
+        // the one thing they can do about it.
+        output.push_str("Waiting on you");
+        let waiting: Vec<_> = self
+            .checks()
+            .iter()
+            .filter(|check| check.outcome() == LintOutcome::Pass && check.action_code().is_some())
+            .collect();
+        append_selected(&mut output, &waiting);
+        output.push_str("Incomplete");
+        let incomplete: Vec<_> = self
+            .checks()
+            .iter()
+            .filter(|check| !matches!(check.outcome(), LintOutcome::Pass | LintOutcome::Finding))
+            .collect();
+        append_selected(&mut output, &incomplete);
+        output
+    }
+}
+
+fn append_findings(
+    output: &mut String,
+    label: &str,
+    report: &LintReport,
+    gate_effect: LintGateEffect,
+) {
+    output.push_str(label);
+    let selected: Vec<_> = report
+        .checks()
+        .iter()
+        .filter(|check| {
+            check.outcome() == LintOutcome::Finding && check.gate_effect() == gate_effect
+        })
+        .collect();
+    append_selected(output, &selected);
+}
+
+fn append_selected(output: &mut String, checks: &[&LintCheckResult]) {
+    if checks.is_empty() {
+        output.push_str(": none\n");
+        return;
+    }
+    output.push_str(&format!(" ({}):\n", checks.len()));
+    for check in checks {
+        let summary = summary_name(check.summary_code());
+        let code_suffix = match (check.recommendation_code(), check.action_code()) {
+            (Some(recommendation), _) => {
+                format!("; recommendation: {}", recommendation_name(recommendation))
+            }
+            (None, Some(action)) => format!("; action: {}", action_name(action)),
+            (None, None) => String::new(),
+        };
+        output.push_str(&format!("  {}: {summary}{code_suffix}\n", check.check_id()));
+        // The codes above are stable identifiers for scripts. This line is the
+        // same thing in plain English for the person reading the terminal.
+        output.push_str(&format!(
+            "    {}{}\n",
+            check.summary_code().meaning(),
+            match (check.recommendation_code(), check.action_code()) {
+                (Some(recommendation), _) => format!(" {}", recommendation.action()),
+                (None, Some(action)) => format!(" {}", action.action()),
+                (None, None) => String::new(),
+            }
+        ));
+        let affected = check.metrics().iter().find_map(|metric| {
+            if metric.code() == LintMetricCode::AffectedRecords {
+                match metric.value() {
+                    LintMetricValue::Count { value } => Some(*value),
+                    LintMetricValue::Boolean { .. } | LintMetricValue::CatalogCode { .. } => None,
+                }
+            } else {
+                None
+            }
+        });
+        let mut evidence_items = check
+            .evidence()
+            .iter()
+            .take(8)
+            .map(evidence_name)
+            .collect::<Vec<_>>();
+        if check.evidence().len() > evidence_items.len() {
+            evidence_items.push(format!(
+                "+{}_more",
+                check.evidence().len() - evidence_items.len()
+            ));
+        }
+        let evidence = evidence_items.join(",");
+        output.push_str(&format!(
+            "    affected={}; evaluated={}/{}; evidence={}; truncated={}\n",
+            affected.map_or_else(|| "unknown".to_string(), |value| value.to_string()),
+            check.coverage().evaluated(),
+            check.coverage().denominator(),
+            if evidence.is_empty() {
+                "none"
+            } else {
+                &evidence
+            },
+            check.coverage().truncated(),
+        ));
+    }
+}
+
+const fn group_index(group: LintCheckGroup) -> usize {
+    match group {
+        LintCheckGroup::Identity => 0,
+        LintCheckGroup::KnowledgeGraph => 1,
+        LintCheckGroup::Memories => 2,
+        LintCheckGroup::Operations => 3,
+        LintCheckGroup::Pages => 4,
+        LintCheckGroup::Runtime => 5,
+        LintCheckGroup::Serving => 6,
+    }
+}
+
+fn evidence_name(evidence: &LintEvidenceRef) -> String {
+    match evidence {
+        LintEvidenceRef::OpaqueId { opaque_id } => format!("opaque:{}", opaque_id.ordinal()),
+        LintEvidenceRef::OpaqueDigest { opaque_digest } => {
+            format!("opaque-digest:{}", opaque_digest.as_str())
+        }
+        LintEvidenceRef::ReasonCode { reason_code } => {
+            format!("reason:{}", reason_name(*reason_code))
+        }
+        LintEvidenceRef::SafeRootRelativePath {
+            safe_root_relative_path,
+        } => format!("path:{safe_root_relative_path:?}"),
+        LintEvidenceRef::SemanticFinding { finding } => format!(
+            "semantic:{}:{:?}:{:?}:{}:{:?}",
+            finding.candidate_id().ordinal(),
+            finding.proposed_action(),
+            finding.reason_code(),
+            finding.confidence_basis_points(),
+            finding.provider_route(),
+        ),
+    }
+}
+
+const fn reason_name(reason: LintReasonCode) -> &'static str {
+    match reason {
+        LintReasonCode::MissingArtifact => "missing_artifact",
+        LintReasonCode::InvalidCatalogState => "invalid_catalog_state",
+        LintReasonCode::ExpectedEmptySubstrate => "expected_empty_substrate",
+        LintReasonCode::InvalidSourceConfiguration => "invalid_source_configuration",
+        LintReasonCode::TerminalOperationFailure => "terminal_operation_failure",
+        LintReasonCode::ExpiredRetry => "expired_retry",
+        LintReasonCode::InvalidOperationState => "invalid_operation_state",
+        LintReasonCode::DurableNoProgress => "durable_no_progress",
+        LintReasonCode::SemanticProviderUnavailable => "semantic_provider_unavailable",
+        LintReasonCode::InsufficientSemanticEvidence => "insufficient_semantic_evidence",
+        LintReasonCode::SemanticExecutionFailure => "semantic_execution_failure",
+        LintReasonCode::SemanticAgentAdjudicationRequired => "semantic_agent_adjudication_required",
+        LintReasonCode::SemanticAgentWorkStale => "semantic_agent_work_stale",
+        LintReasonCode::SemanticAgentSubmissionInvalid => "semantic_agent_submission_invalid",
+        LintReasonCode::SemanticCandidateGenerationFailure => {
+            "semantic_candidate_generation_failure"
+        }
+        LintReasonCode::SemanticPopulationIncomplete => "semantic_population_incomplete",
+        LintReasonCode::SemanticDisagreementUnresolved => "semantic_disagreement_unresolved",
+        LintReasonCode::SemanticSecondJudgeRequired => "semantic_second_judge_required",
+    }
+}
+
+const fn recommendation_name(recommendation: LintRecommendationCode) -> &'static str {
+    match recommendation {
+        LintRecommendationCode::ReviewFinding => "review_finding",
+        LintRecommendationCode::RestorePrerequisite => "restore_prerequisite",
+        LintRecommendationCode::RerunAfterSnapshotStabilizes => "rerun_after_snapshot_stabilizes",
+        LintRecommendationCode::InspectRuntime => "inspect_runtime",
+    }
+}
+
+const fn action_name(action: LintActionCode) -> &'static str {
+    match action {
+        LintActionCode::ChooseModelSource => "choose_model_source",
+    }
+}
+
+const fn summary_name(summary: LintSummaryCode) -> &'static str {
+    match summary {
+        LintSummaryCode::CheckPassed => "check_passed",
+        LintSummaryCode::FindingDetected => "finding_detected",
+        LintSummaryCode::PrerequisiteUnavailable => "prerequisite_unavailable",
+        LintSummaryCode::SnapshotInconsistent => "snapshot_inconsistent",
+        LintSummaryCode::ExecutionFailed => "execution_failed",
+        LintSummaryCode::ExpectedEmpty => "expected_empty",
+    }
+}
